requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
pydantic==2.10.6
pydantic-settings==2.7.1
SQLAlchemy==2.0.38
asyncpg==0.30.0
alembic==1.14.1
psycopg[binary]==3.2.4
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
httpx==0.27.2
redis==5.2.1
celery==5.4.0
google-auth==2.38.0
python-dotenv==1.0.1
pytesseract==0.3.13
Pillow==11.1.0
reportlab==4.2.5
pytest==8.3.4
pytest-asyncio==0.25.3
